add atari gamepad
