--- BKP/ps2in.h
@@ -1,42 +1,0 @@
-/*
- * The MIT License (MIT)
- *
- * Copyright (c) 2024 No0ne (https://github.com/No0ne)
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- *
- */
-
-
-#include "hardware/pio.h"
-#include "ps2x2pico.h"
-
-typedef struct {
-  PIO pio;
-  uint sm;
-  u8 state;
-  u8 byte;
-} ps2in;
-
-
-void ps2in_init(ps2in* this, PIO pio, u8 data_pin);
-void ps2in_task(ps2in* this, ps2out* out);
-void ps2in_reset(ps2in* this);
-void ps2in_set(ps2in* this, u8 command, u8 byte);
-void ms_send_movement(u8 buttons, s8 x, s8 y, s8 z);
