--- BKP/ps2out.h
@@ -1,48 +1,0 @@
-/*
- * The MIT License (MIT)
- *
- * Copyright (c) 2024 No0ne (https://github.com/No0ne)
- *           (c) 2023 Dustin Hoffman
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- *
- */
-
-
-#include "hardware/pio.h"
-#include "pico/util/queue.h"
-#include "ps2x2pico.h"
-
-u32 ps2_frame(u8 byte);
-typedef void (*rx_callback)(u8 byte, u8 prev_byte);
-
-typedef struct {
-  PIO pio;
-  uint sm;
-  queue_t qbytes;
-  queue_t qpacks;
-  rx_callback rx;
-  u8 last_rx;
-  u8 last_tx;
-  u8 sent;
-  u8 busy;
-} ps2out;
-
-void ps2out_init(ps2out* this, PIO pio, u8 data_pin, rx_callback rx);
-void ps2out_task(ps2out* this);
